requirements.txt
diffusers==0.15.0
transformers==4.27.4
streamlit==1.21.0
pillow==8.4.0
numpy==1.19.5
